rename(apps/app): 공통 컴포넌트 폴더 이동
