+++ apps/app/src/components/@common/Checkbox/Checkbox.stories.tsx
