+++ apps/app/src/components/@common/Icon.tsx
@@ -1,7 +1,7 @@
 import {CSSProperties, forwardRef, Ref} from 'react';
 import {KeyOfPalette, theme} from '@greeenai/design-tokens';
-import {KeyOfIcons} from '../types/Icon';
-import * as Icons from '../assets/icons';
+import {KeyOfIcons} from '../../types/Icon';
+import * as Icons from '../../assets/icons';
 import {StyleProp, View, ViewStyle} from 'react-native';
 
 type IconProps = {
@@ -9,7 +9,7 @@ type IconProps = {
   width?: CSSProperties['width'];
   height?: CSSProperties['height'];
   fill?: KeyOfPalette;
-  style?: StyleProp<ViewStyle>
+  style?: StyleProp<ViewStyle>;
   onPress?: (...args: any[]) => void;
 };
